app/requirements.txt
django
djangorestframework
djangorestframework-gis
django-rest-swagger
django-cors-headers
psycopg2-binary
pytest
pytest-django
shapely
awsebcli
